F1-53: Update main window & Setup import/ export panel
- The home page view and VM were removed because they
are not necessary when using windows; we can just use the
main window view and VM instead.

- I have created a basic main window view to display what
options should be visible from the main pane:
I have added the record/ replay buttons too ready for testing
the other windows when they are created.

- The warnings are now displayed differently: CheckWarnings()
now does the brunt of it by setting the Warning. If a warningMessage
exists, it will be displayed and the Start/Stop button will be disabled.
- If Replay is checked and a file isn't selected, it will display as a
warning.
- If the version selected is not supported, that warning will take
precedence even if the replay warning is also true.